formatted code with "black" formatter

+++ zahner_analysis/file_import/thales_file_utils.py
@@ -28,22 +28,28 @@ import io
 import struct
 import datetime
 
+
 def readF8FromFile(file):
     bytesRead = file.read(8)
-    retval = struct.unpack('>d', bytesRead)
+    retval = struct.unpack(">d", bytesRead)
     return retval[0]
 
+
 def readF8ArrayFromFile(file, length):
-    return np.ndarray(shape=(length,), dtype='>f8', buffer=file.read(8 * length))
+    return np.ndarray(shape=(length,), dtype=">f8", buffer=file.read(8 * length))
+
 
 def readI2FromFile(file):
     return int.from_bytes(file.read(2), "big", signed=True)
 
+
 def readI2ArrayFromFile(file, length):
-    return np.ndarray(shape=(length,), dtype='>i2', buffer=file.read(2 * length))
+    return np.ndarray(shape=(length,), dtype=">i2", buffer=file.read(2 * length))
+
 
 def readI6FromFile(file):
     return int.from_bytes(file.read(6), "big", signed=True)
+
 
 def readTimeStampDateTimeArrayFromFile(file, length):
     timeArray = readF8ArrayFromFile(file, length)
@@ -52,23 +58,25 @@ def readTimeStampDateTimeArrayFromFile(file, length):
         datetimeArray.append(thalesTimeStampToDateTime(time))
     return datetimeArray
 
+
 def readZahnerStringFromFile(file):
     length = readI2FromFile(file)
     content = bytearray()
-    
+
     for i in range(length):
-        content.append(file.read(1)[0])         
-    
+        content.append(file.read(1)[0])
+
     return content.decode("ASCII").swapcase()
+
 
 def readZahnerDate(file):
     dateString = readZahnerStringFromFile(file)
     date = dateString[0:6]
-    
+
     day = int(date[0:2])
     month = int(date[2:4])
     year = int(date[4:6])
-    
+
     """
     Only the last two digits of the date are saved.
     It is assumed that the measurement was carried out between 1970 and 2070.
@@ -78,21 +86,20 @@ def readZahnerDate(file):
         year += 2000
     else:
         year += 1900
-        
+
     return datetime.datetime(year, month, day)
 
+
 def thalesTimeStampToDateTime(timestamp):
-    """ Calculation of the time stamp.
-    
+    """Calculation of the time stamp.
+
     The time is in seconds related to 01.01.1980.
-    
+
     :param timestamp: Seconds since 01.01.1980.
     :returns: Python datetime object.
     """
     timeZero = datetime.datetime(1980, 1, 1)
     timeDifference = datetime.timedelta(seconds=abs(timestamp))
-    
+
     timestamp = timeZero + timeDifference
     return timestamp
-    
-
